requirements.txt
transformers==4.44.2
torch==2.4.0
sentencepiece==0.2.0
protobuf==5.28.0
selenium
webdriver-manager
google-play-scraper 
app-store-scraper
langchain 
langchain_groq
groq
openai==0.28
python-dotenv